+++ localflavor/ee/forms.py
@@ -1,9 +1,9 @@
 import re
 from datetime import date
 
-from django.core.validators import EMPTY_VALUES
+from django.core.exceptions import ImproperlyConfigured
 from django.forms import ValidationError
-from django.forms.fields import Field, RegexField, Select
+from django.forms.fields import CharField, RegexField, Select
 from django.utils.translation import gettext_lazy as _
 
 from .ee_counties import COUNTY_CHOICES
@@ -35,7 +35,7 @@ class EECountySelect(Select):
         super().__init__(attrs, choices=COUNTY_CHOICES)
 
 
-class EEPersonalIdentificationCode(Field):
+class EEPersonalIdentificationCode(CharField):
     """A form field that validates input as an Estonian personal identification code.
 
     See: https://www.riigiteataja.ee/akt/106032012004
@@ -45,6 +45,11 @@ class EEPersonalIdentificationCode(Field):
         'invalid_format': _('Enter an 11-digit Estonian personal identification code.'),
         'invalid': _('Enter a valid Estonian personal identification code.'),
     }
+
+    def __init__(self, **kwargs):
+        if "strip" in kwargs and kwargs["strip"] is False:
+            raise ImproperlyConfigured("strip cannot be set to False")
+        super().__init__(**kwargs)
 
     @staticmethod
     def ee_checksum(value):
@@ -63,8 +68,8 @@ class EEPersonalIdentificationCode(Field):
 
     def clean(self, value):
         value = super().clean(value)
-        if value in EMPTY_VALUES:
-            return ''
+        if value in self.empty_values:
+            return value
 
         match = re.match(idcode, value)
         if not match:
@@ -89,7 +94,7 @@ class EEPersonalIdentificationCode(Field):
         return value
 
 
-class EEBusinessRegistryCode(Field):
+class EEBusinessRegistryCode(CharField):
     """A form field that validates input as an Estonian business registration code.
 
     .. versionadded:: 1.2
@@ -100,11 +105,15 @@ class EEBusinessRegistryCode(Field):
         'invalid': _('Enter a valid Estonian business registry code.'),
     }
 
+    def __init__(self, **kwargs):
+        if "strip" in kwargs and kwargs["strip"] is False:
+            raise ImproperlyConfigured("strip cannot be set to False")
+        super().__init__(**kwargs)
+
     def clean(self, value):
         value = super().clean(value)
-        if value in EMPTY_VALUES:
-            return ''
-        value = value.strip()
+        if value in self.empty_values:
+            return value
 
         match = re.match(bregcode, value)
         if not match: